# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pizza  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "View all articles",
                "description": "List of all published articles",
                "operationId": "App\\Http\\Api\\Controllers\\ArticleApiController::index",
                "responses": {
                    "200": {
                        "description": "Succes response with all published articles"
                    },
                    "204": {
                        "description": "Succes response no published articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No published articles"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/article-category/{categoryId}": {
            "get": {
                "tags": [
                    "Article Category"
                ],
                "summary": "Articles by category",
                "description": "Show published articles by category",
                "operationId": "App\\Http\\Api\\Controllers\\ArticleCategoryApiController::show",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succes show existing published article from category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "404": {
                "description": "Errore, no suche category",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Sorry, wrong email address or password. Please try again"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}